+++ ft_strlcpy.c
@@ -1,7 +1,7 @@
-unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int dsize)
+size_t	ft_strlcpy(char *dst, const char *src, size_t dsize)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
